+++ src/etc/di.php
@@ -1,7 +1,8 @@
 <?php
 declare(strict_types=1);
 
-use Vconnect\IntegrityChecker\Analysis\Service\Dependencies\Scanner\{DbSchema,
+use Vconnect\IntegrityChecker\Analysis\Service\Dependencies\Scanner\{DbDDL,
+    DbSchema,
     GraphQlSchema,
     PhpFiles,
     QueueConfig,
@@ -15,6 +16,7 @@ return [
             DI\get(XmlConfigFiles::class),
             DI\get(DbSchema::class),
             DI\get(QueueConfig::class),
-            DI\get(GraphQlSchema::class)
+            DI\get(GraphQlSchema::class),
+            DI\get(DbDDL::class),
         ])
 ];